Pulled spawning code from main into controller.spawner
controller.spawner will be the brains for spawning creeps. hopefully
bully out hardcoded minimums later on.
