+++ main.js
@@ -8,6 +8,8 @@ var roleBuilder = require('role.builder');
 var roleRepairer = require('role.repairer');
 var roleWallRepairer = require('role.wallRepairer');
 var assignSource = require('assignSource');
+
+var creepSpawner = require('controller.spawner');
 //nochange
 
 module.exports.loop = function () {
@@ -22,28 +24,28 @@ module.exports.loop = function () {
         }
     }
     
-    var ramparts =_.filter(Game.structures, {structureType: STRUCTURE_RAMPART});
-    var ramp;
+    //var ramparts =_.filter(Game.structures, {structureType: STRUCTURE_RAMPART});
+    //var ramp;
     
-    for(var roomName in Game.rooms)
-    {
-        var room = Game.rooms[roomName];
-        room.memory.ramparts = ramparts;
-        for(var i in ramparts)
-        {
-            var ramp = ramparts[i];
-            if(!ramp.memory)
-            {   
-                ramp.memory = {};
-                ramp.memory.full = {};
-                ramp.memory.full = false;
-            }
-        }   
-    }
+    //for(var roomName in Game.rooms)
+    //{
+    //    var room = Game.rooms[roomName];
+    //    room.memory.ramparts = ramparts;
+    //    for(var i in ramparts)
+    //    {
+    //        var ramp = ramparts[i];
+    //        if(!ramp.memory)
+    //        {   
+    //            ramp.memory = {};
+    //            ramp.memory.full = {};
+    //            ramp.memory.full = false;
+    //        }
+    //    }   
+    //}
     
     
 
-    var harvesters = _.filter(Game.creeps, {memory: 'harvester'});
+    //var harvesters = _.filter(Game.creeps, {memory: 'harvester'});
     //assignSource.run(harvesters);
     
     // for every creep name in Game.creeps
@@ -76,6 +78,7 @@ module.exports.loop = function () {
         }
     }
 
+    
     //var towers = Game.rooms.find(FIND_STRUCTURES, {
     //    filter: (s) => s.structureType == STRUCTURE_TOWER
     //});
@@ -86,70 +89,5 @@ module.exports.loop = function () {
         }
     }*/
 
-    // setup some minimum numbers for different roles
-    var minimumNumberOfHarvesters = 4;
-    var minimumNumberOfUpgraders = 3;
-    var minimumNumberOfBuilders = 3;
-    var minimumNumberOfRepairers = 1;
-    var minimumNumberOfWallRepairers = 1;
-    var minimumNumberOfDefenders= 4;
-    
-    // count the number of creeps alive for each role
-    // _.sum will count the number of properties in Game.creeps filtered by the
-    //  arrow function, which checks for the creep being a harvester
-    var numberOfHarvesters = _.sum(Game.creeps, (c) => c.memory.role == 'harvester');
-    var numberOfUpgraders = _.sum(Game.creeps, (c) => c.memory.role == 'upgrader');
-    var numberOfBuilders = _.sum(Game.creeps, (c) => c.memory.role == 'builder');
-    var numberOfRepairers = _.sum(Game.creeps, (c) => c.memory.role == 'repairer');
-    var numberOfWallRepairers = _.sum(Game.creeps, (c) => c.memory.role == 'wallRepairer');
-    var numberOfDefenders = _.sum(Game.creeps, (c) => c.memory.role =='defender');
-
-    var energy = Game.spawns.Alpha.room.energyCapacityAvailable;
-    var name = undefined;
-
-    // if not enough harvesters
-    if (numberOfHarvesters < minimumNumberOfHarvesters) {
-        // try to spawn one
-        name = Game.spawns.Alpha.createCustomCreep(energy, 'harvester');
-
-        // if spawning failed and we have no harvesters left
-        if (name == ERR_NOT_ENOUGH_ENERGY && numberOfHarvesters == 0) {
-            // spawn one with what is available
-            name = Game.spawns.Alpha.createCustomCreep(
-                Game.spawns.Alpha.room.energyAvailable, 'harvester');
-        }
-    }
-    // if not enough upgraders
-    else if (numberOfUpgraders < minimumNumberOfUpgraders) {
-        // try to spawn one
-        name = Game.spawns.Alpha.createCustomCreep(energy, 'upgrader');
-    }
-    // if not enough repairers
-    else if (numberOfRepairers < minimumNumberOfRepairers) {
-        // try to spawn one
-        name = Game.spawns.Alpha.createCustomCreep(energy, 'repairer');
-    }
-    // if not enough builders
-    else if (numberOfBuilders < minimumNumberOfBuilders) {
-        // try to spawn one
-        name = Game.spawns.Alpha.createCustomCreep(energy, 'builder');
-    }
-    // if not enough wallRepairers
-    else if (numberOfWallRepairers < minimumNumberOfWallRepairers) {
-        // try to spawn one
-        name = Game.spawns.Alpha.createCustomCreep(energy, 'wallRepairer');
-    }
-    else if(numberOfDefenders < minimumNumberOfDefenders){
-        name = Game.spawns.Alpha.createMilitaryCreep(energy,'defender');
-    }
-    else {
-        // else try to spawn a builder
-        name = Game.spawns.Alpha.createCustomCreep(energy, 'builder');
-    }
-    
-    // print name to console if spawning was a success
-    // name > 0 would not work since string > 0 returns false
-    if (!(name < 0)) {
-        console.log("Spawned new creep: " + name);
-    }
+    creepSpawner.run();
 };
